requirements.txt
python-dotenv==1.2.2
pyTelegramBotAPI==4.31.0
langchain==1.2.10
langchain_openai==1.1.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-classic==1.0.1
langchain-core==1.2.18
langchain-text-splitters==1.1.1
chromadb==1.5.2
sentence-transformers==5.2.3
#pypdf==6.7.5
docling==2.80.0
tavily-python==0.7.22
arxiv==2.4.1
semanticscholar==0.11.0